endpointsynchronizer: move from pkg/k8s/watchers to pkg/endpointmanager
Moves endpointsynchronizer from pkg/k8s/watchers to pkg/endpointmanager
so that endpointmanager does not need to import top-level pkg/k8s/watchers.
This avoids import loops for endpointmanager dependants.

Suggested-by: Jussi Maki <jussi@isovalent.com>
