+++ pkg/endpointmanager/cell.go
@@ -16,7 +16,6 @@ import (
 	"github.com/cilium/cilium/pkg/identity/cache"
 	"github.com/cilium/cilium/pkg/ipcache"
 	"github.com/cilium/cilium/pkg/k8s/client"
-	"github.com/cilium/cilium/pkg/k8s/watchers"
 	"github.com/cilium/cilium/pkg/k8s/watchers/subscriber"
 	"github.com/cilium/cilium/pkg/metrics"
 	"github.com/cilium/cilium/pkg/option"
@@ -33,6 +32,7 @@ var Cell = cell.Module(
 
 	cell.Config(defaultEndpointManagerConfig),
 	cell.Provide(newDefaultEndpointManager),
+	cell.ProvidePrivate(newEndpointSynchronizer),
 )
 
 type EndpointsLookup interface {
@@ -175,6 +175,13 @@ type EndpointManager interface {
 	CallbackForEndpointsAtPolicyRev(ctx context.Context, rev uint64, done func(time.Time)) error
 }
 
+// EndpointResourceSynchronizer is an interface which synchronizes CiliumEndpoint
+// resources with Kubernetes.
+type EndpointResourceSynchronizer interface {
+	RunK8sCiliumEndpointSync(ep *endpoint.Endpoint, conf endpoint.EndpointStatusConfiguration, hr cell.HealthReporter)
+	DeleteK8sCiliumEndpointSync(e *endpoint.Endpoint)
+}
+
 var (
 	_ EndpointsLookup = &endpointManager{}
 	_ EndpointsModify = &endpointManager{}
@@ -189,6 +196,7 @@ type endpointManagerParams struct {
 	Clientset       client.Clientset
 	MetricsRegistry *metrics.Registry
 	Scope           cell.Scope
+	EPSynchronizer  EndpointResourceSynchronizer
 }
 
 type endpointManagerOut struct {
@@ -202,7 +210,7 @@ type endpointManagerOut struct {
 func newDefaultEndpointManager(p endpointManagerParams) endpointManagerOut {
 	checker := endpoint.CheckHealth
 
-	mgr := New(&watchers.EndpointSynchronizer{Clientset: p.Clientset}, p.Scope)
+	mgr := New(p.EPSynchronizer, p.Scope)
 	if p.Config.EndpointGCInterval > 0 {
 		ctx, cancel := context.WithCancel(context.Background())
 		p.Lifecycle.Append(hive.Hook{
@@ -225,4 +233,14 @@ func newDefaultEndpointManager(p endpointManagerParams) endpointManagerOut {
 		Modify:  mgr,
 		Manager: mgr,
 	}
+}
+
+type endpointSynchronizerParams struct {
+	cell.In
+
+	Clientset client.Clientset
+}
+
+func newEndpointSynchronizer(p endpointSynchronizerParams) EndpointResourceSynchronizer {
+	return &EndpointSynchronizer{Clientset: p.Clientset}
 }
